HTTPステータスコードが200番以外の場合それ以降の動作をスキップ

+++ delayInfo.rb
@@ -6,8 +6,14 @@ require 'sqlite3'
 URL = 'https://tetsudo.rti-giken.jp/free/delay.json'
 uri = URI.parse(URL)
 
-json = Net::HTTP.get(uri)
-result = JSON.parse(json)
+response = Net::HTTP.get_response(uri)
+
+unless response.code == '200' 
+  puts 'Failed to get JSON'
+  exit
+end
+
+result = JSON.parse(response.body)
 
 db = SQLite3::Database.new 'delayInfo.db'
 db.execute('DELETE FROM infoTable')
@@ -22,4 +28,5 @@ result.each do |data|
   SQL
   db.execute(sql,data["name"],data["company"],data["lastupdate_gmt"].to_i)
 end
+
 puts 'success'
